pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "GrandEntrance"
description = "Grand Entrance Style Audio Player."
authors = [{name = "Joe Stanley", email = "engineerjoe440@yahoo.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
dependencies = [
    "PySimpleGUI",
    "python-vlc",
]
dynamic = ["version"]

[project.gui-scripts]
song-barn = "GrandEntrance:entrypoint"
